requirements.txt
django
djangorestframework
click
requests
